Move non-Superset-class views out of views/core.py
